+++ ntex-io/src/lib.rs
@@ -16,7 +16,7 @@ mod ioref;
 mod seal;
 mod tasks;
 mod timer;
-pub mod utils;
+mod utils;
 
 use ntex_bytes::BytesMut;
 use ntex_codec::{Decoder, Encoder};
@@ -24,10 +24,11 @@ use ntex_util::time::Millis;
 
 pub use self::dispatcher::Dispatcher;
 pub use self::filter::Base;
-pub use self::io::{Io, IoRef};
+pub use self::framed::Framed;
+pub use self::io::{Io, IoRef, OnDisconnect};
 pub use self::seal::{IoBoxed, Sealed};
 pub use self::tasks::{ReadContext, WriteContext};
-pub use self::utils::filter;
+pub use self::utils::{filter, seal};
 
 /// Status for read task
 #[derive(Copy, Clone, Debug, PartialEq, Eq, Hash)]