+++ ntex-io/src/utils.rs
@@ -1,10 +1,8 @@
-use std::{future::Future, marker::PhantomData, pin::Pin, task::Context, task::Poll};
+use std::{marker::PhantomData, task::Context, task::Poll};
 
 use ntex_service::{fn_factory_with_config, into_service, Service, ServiceFactory};
-use ntex_util::{future::Ready, ready};
+use ntex_util::future::Ready;
 
-pub use crate::framed::Framed;
-pub use crate::io::OnDisconnect;
 use crate::{Filter, FilterFactory, Io, IoBoxed};
 
 /// Service that converts any Io<F> stream to IoBoxed stream
@@ -30,18 +28,6 @@ where
     })
 }
 
-/// Service that converts Io<F> responses from service to the IoBoxed
-pub fn boxed<S, R, F>(inner: S) -> Boxed<S, R>
-where
-    F: Filter,
-    S: Service<R, Response = Io<F>>,
-{
-    Boxed {
-        inner,
-        _t: PhantomData,
-    }
-}
-
 /// Create filter factory service
 pub fn filter<T, F>(filter: T) -> FilterServiceFactory<T, F>
 where
@@ -51,126 +37,6 @@ where
     FilterServiceFactory {
         filter,
         _t: PhantomData,
-    }
-}
-
-pub struct BoxedFactory<S, R> {
-    inner: S,
-    _t: PhantomData<R>,
-}
-
-impl<S, R> BoxedFactory<S, R> {
-    pub fn new(inner: S) -> Self {
-        Self {
-            inner,
-            _t: PhantomData,
-        }
-    }
-}
-
-impl<S: Clone, R> Clone for BoxedFactory<S, R> {
-    fn clone(&self) -> Self {
-        Self::new(self.inner.clone())
-    }
-}
-
-impl<S, R, C, F> ServiceFactory<R, C> for BoxedFactory<S, R>
-where
-    F: Filter,
-    S: ServiceFactory<R, C, Response = Io<F>>,
-{
-    type Response = IoBoxed;
-    type Error = S::Error;
-    type Service = Boxed<S::Service, R>;
-    type InitError = S::InitError;
-    type Future = BoxedFactoryResponse<S, R, C>;
-
-    fn new_service(&self, cfg: C) -> Self::Future {
-        BoxedFactoryResponse {
-            fut: self.inner.new_service(cfg),
-            _t: PhantomData,
-        }
-    }
-}
-
-pub struct Boxed<S, R> {
-    inner: S,
-    _t: PhantomData<R>,
-}
-
-impl<S, R> Boxed<S, R> {
-    pub fn new(inner: S) -> Self {
-        Self {
-            inner,
-            _t: PhantomData,
-        }
-    }
-}
-
-impl<S: Clone, R> Clone for Boxed<S, R> {
-    fn clone(&self) -> Self {
-        Self::new(self.inner.clone())
-    }
-}
-
-impl<S, R, F> Service<R> for Boxed<S, R>
-where
-    F: Filter,
-    S: Service<R, Response = Io<F>>,
-{
-    type Response = IoBoxed;
-    type Error = S::Error;
-    type Future = BoxedResponse<S, R>;
-
-    fn poll_ready(&self, cx: &mut Context<'_>) -> Poll<Result<(), S::Error>> {
-        self.inner.poll_ready(cx)
-    }
-
-    #[inline]
-    fn poll_shutdown(&self, cx: &mut Context<'_>, is_err: bool) -> Poll<()> {
-        self.inner.poll_shutdown(cx, is_err)
-    }
-
-    fn call(&self, req: R) -> Self::Future {
-        BoxedResponse {
-            fut: self.inner.call(req),
-        }
-    }
-}
-
-pin_project_lite::pin_project! {
-    #[doc(hidden)]
-    pub struct BoxedFactoryResponse<S: ServiceFactory<R, C>, R, C> {
-        #[pin]
-        fut: S::Future,
-        _t: PhantomData<(R, C)>
-    }
-}
-
-impl<S: ServiceFactory<R, C>, R, C> Future for BoxedFactoryResponse<S, R, C> {
-    type Output = Result<Boxed<S::Service, R>, S::InitError>;
-
-    fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-        Poll::Ready(ready!(self.project().fut.poll(cx)).map(|inner| Boxed {
-            inner,
-            _t: PhantomData,
-        }))
-    }
-}
-
-pin_project_lite::pin_project! {
-    #[doc(hidden)]
-    pub struct BoxedResponse<S: Service<R>, R> {
-        #[pin]
-        fut: S::Future,
-    }
-}
-
-impl<S: Service<R, Response = Io<F>>, R, F: Filter> Future for BoxedResponse<S, R> {
-    type Output = Result<IoBoxed, S::Error>;
-
-    fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-        Poll::Ready(ready!(self.project().fut.poll(cx)).map(IoBoxed::from))
     }
 }